requirements.txt
# Core dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
urllib3>=2.0.0
pytest>=7.3.1

# LangChain ecosystem
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-callbacks>=0.0.2

# Embeddings and LLMs
openai>=1.2.0

# Web scraping tools
firecrawl>=0.1.0

# Data processing
pandas>=2.0.0

# Optional dependencies
nltk>=3.8.0

# Development dependencies
black>=23.3.0
flake8>=6.0.0